refactor: rename test folder to tests
